HUM-92 - code refactor
