+++ packages/apps/human-app/frontend/src/modules/worker/hooks/use-register-address/index.ts
@@ -1,1 +1,3 @@
 export { useRegisterAddress } from './use-register-address';
+export { useRegisterAddressMutation } from './use-register-address-mutation';
+export { useRegisterAddressValidation } from './use-register-address-validation';
