--- packages/apps/human-app/frontend/src/modules/worker/hooks/use-register-address/types.ts
@@ -1,15 +1,0 @@
-import type { ResponseError } from '@/shared/types/global.type';
-
-export interface RegisterAddressCallbacks {
-  onSuccess?: () => void | Promise<void>;
-  onError?: (error: ResponseError) => void | Promise<void>;
-}
-
-export interface RegisterAddressParams {
-  address: string;
-  signature: string;
-}
-
-export interface RegisterAddressResponse {
-  success: boolean;
-}
